website/requirement.txt
certifi==2021.10.8
charset-normalizer==2.0.9
Django==2.2
django-ckeditor==6.2.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
martor==1.6.4
mysqlclient==2.1.0
protobuf==3.17.3
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
zipp==3.6.0
